requirements.txt
google-generativeai==0.7.1
Pillow==10.4.0
python-dotenv==1.0.1
Flask==3.0.3
Flask-Cors==4.0.1
pydantic==2.5.0
gunicorn==21.2.0
requests==2.31.0
